refactor(onboard): extract gateway HTTP readiness helpers to src/lib/onboard/
